unexpand nodes, even with blank nodes expanded
