+++ src/shared/containers/GraphContainer/Graph.ts
@@ -51,6 +51,7 @@ export const createNodesAndEdgesFromResourceLinks = (
               id: `edge-${originId}-${link['@id']}`,
               source: originId,
               target: link['@id'],
+              parentId: originId,
             },
           },
         ];
@@ -75,7 +76,7 @@ export const createNodesAndEdgesFromResourceLinks = (
               id: `edge-${originId}-${link['@id']}`,
               source: originId,
               target: link['@id'],
-              parentId: link['@id'],
+              parentId: originId,
             },
           };
         }
@@ -86,7 +87,7 @@ export const createNodesAndEdgesFromResourceLinks = (
               id: `edge-${originId}-${blankNode && blankNode.data.id}`,
               source: originId,
               target: blankNode && blankNode.data.id,
-              parentId: link['@id'],
+              parentId: originId,
             },
           };
         }
@@ -98,7 +99,7 @@ export const createNodesAndEdgesFromResourceLinks = (
               id: `edge-${prev && prev.data.id}-${link['@id']}`,
               source: prev && prev.data.id,
               target: link['@id'],
-              parentId: link['@id'],
+              parentId: originId,
             },
           };
         }
@@ -109,7 +110,7 @@ export const createNodesAndEdgesFromResourceLinks = (
               blankNode.data.id}`,
             source: prev && prev.data.id,
             target: blankNode && blankNode.data.id,
-            parentId: link['@id'],
+            parentId: originId,
           },
         };
       });
@@ -129,7 +130,38 @@ export const makeBlankNodes = (
     data: {
       id: `${resourceId}-${path}-${linkId}`,
       isBlankNode: true,
-      parentId: linkId,
+      parentId: resourceId,
     },
   };
+};
+
+export const getListOfChildrenRecursive = (
+  parentId: string,
+  elements: cytoscape.ElementDefinition[]
+): string[] => {
+  const targetNode = elements.find(element => element.data.id === parentId);
+  if (!targetNode) {
+    return [];
+  }
+  targetNode.data.isExpanded = false;
+
+  const childIds = elements
+    .filter(
+      element =>
+        element.data.source === parentId || element.data.parentId === parentId
+    )
+    .map(child => child.data.id)
+    .filter(Boolean) as string[];
+
+  const childRecursiveIds = childIds.reduce(
+    (childRecursiveIdsList: string[], id: string) => {
+      return [
+        ...childRecursiveIdsList,
+        ...getListOfChildrenRecursive(id, elements),
+      ];
+    },
+    []
+  );
+
+  return [...childIds, ...childRecursiveIds];
 };
